+++ src/OpenTelemetry.Instrumentation.Wcf/Implementation/InstrumentedDuplexChannel.cs
@@ -7,7 +7,7 @@ using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Instrumentation.Wcf.Implementation;
 
-internal sealed class InstrumentedDuplexChannel : InstrumentedChannel<IDuplexChannel>, IDuplexSessionChannel
+internal class InstrumentedDuplexChannel : InstrumentedChannel<IDuplexChannel>, IDuplexChannel
 {
     private readonly TimeSpan telemetryTimeout;
 
@@ -22,8 +22,6 @@ internal sealed class InstrumentedDuplexChannel : InstrumentedChannel<IDuplexCha
     public EndpointAddress RemoteAddress => this.Inner.RemoteAddress;
 
     public Uri Via => this.Inner.Via;
-
-    public IDuplexSession Session => ((IDuplexSessionChannel)this.Inner).Session;
 
     public void Send(Message message)
     {