+++ test/OpenTelemetry.Instrumentation.GrpcCore.Tests/GrpcCoreServerInterceptorTests.cs
@@ -19,7 +19,7 @@ public class GrpcCoreServerInterceptorTests
     /// <returns>A task.</returns>
     [Fact]
     public async Task UnaryServerHandlerSuccess() =>
-        await TestHandlerSuccess(FoobarService.MakeUnaryAsyncRequest);
+        await TestHandlerSuccess(FoobarService.MakeUnaryAsyncRequest, FoobarService.UnaryMethod);
 
     /// <summary>
     /// Validates a failed UnaryServerHandler call.
@@ -27,7 +27,7 @@ public class GrpcCoreServerInterceptorTests
     /// <returns>A task.</returns>
     [Fact]
     public async Task UnaryServerHandlerFail() =>
-        await TestHandlerFailure(FoobarService.MakeUnaryAsyncRequest);
+        await TestHandlerFailure(FoobarService.MakeUnaryAsyncRequest, FoobarService.UnaryMethod);
 
     /// <summary>
     /// Validates a successful ClientStreamingServerHandler call.
@@ -35,7 +35,7 @@ public class GrpcCoreServerInterceptorTests
     /// <returns>A task.</returns>
     [Fact]
     public async Task ClientStreamingServerHandlerSuccess() =>
-        await TestHandlerSuccess(FoobarService.MakeClientStreamingRequest);
+        await TestHandlerSuccess(FoobarService.MakeClientStreamingRequest, FoobarService.ClientStreamingMethod);
 
     /// <summary>
     /// Validates a failed ClientStreamingServerHandler call.
@@ -43,7 +43,7 @@ public class GrpcCoreServerInterceptorTests
     /// <returns>A task.</returns>
     [Fact]
     public async Task ClientStreamingServerHandlerFail() =>
-        await TestHandlerFailure(FoobarService.MakeClientStreamingRequest);
+        await TestHandlerFailure(FoobarService.MakeClientStreamingRequest, FoobarService.ClientStreamingMethod);
 
     /// <summary>
     /// Validates a successful ServerStreamingServerHandler call.
@@ -51,7 +51,7 @@ public class GrpcCoreServerInterceptorTests
     /// <returns>A task.</returns>
     [Fact]
     public async Task ServerStreamingServerHandlerSuccess() =>
-        await TestHandlerSuccess(FoobarService.MakeServerStreamingRequest);
+        await TestHandlerSuccess(FoobarService.MakeServerStreamingRequest, FoobarService.ServerStreamingMethod);
 
     /// <summary>
     /// Validates a failed ServerStreamingServerHandler call.
@@ -59,7 +59,7 @@ public class GrpcCoreServerInterceptorTests
     /// <returns>A task.</returns>
     [Fact]
     public async Task ServerStreamingServerHandlerFail() =>
-        await TestHandlerFailure(FoobarService.MakeServerStreamingRequest);
+        await TestHandlerFailure(FoobarService.MakeServerStreamingRequest, FoobarService.ServerStreamingMethod);
 
     /// <summary>
     /// Validates a successful DuplexStreamingServerHandler call.
@@ -67,7 +67,7 @@ public class GrpcCoreServerInterceptorTests
     /// <returns>A task.</returns>
     [Fact]
     public async Task DuplexStreamingServerHandlerSuccess() =>
-        await TestHandlerSuccess(FoobarService.MakeDuplexStreamingRequest);
+        await TestHandlerSuccess(FoobarService.MakeDuplexStreamingRequest, FoobarService.DuplexStreamingMethod);
 
     /// <summary>
     /// Validates a failed DuplexStreamingServerHandler call.
@@ -75,7 +75,7 @@ public class GrpcCoreServerInterceptorTests
     /// <returns>A task.</returns>
     [Fact]
     public async Task DuplexStreamingServerHandlerFail() =>
-        await TestHandlerFailure(FoobarService.MakeDuplexStreamingRequest);
+        await TestHandlerFailure(FoobarService.MakeDuplexStreamingRequest, FoobarService.DuplexStreamingMethod);
 
     /// <summary>
     /// Validates that non protobuf payloads do not abort server RPCs when
@@ -111,7 +111,7 @@ public class GrpcCoreServerInterceptorTests
         };
 
         server.Start();
-        var serverUriString = new Uri("dns:localhost:" + server.Ports.Single().BoundPort).ToString();
+        var serverUriString = $"dns:localhost:{server.Ports.Single().BoundPort}";
 
         try
         {
@@ -129,7 +129,9 @@ public class GrpcCoreServerInterceptorTests
                 Assert.NotNull(response);
 
                 var activity = activityListener.Activity;
-                GrpcCoreClientInterceptorTests.ValidateCommonActivityTags(activity);
+
+                GrpcCoreClientInterceptorTests.ValidateCommonActivityTags(activity, FoobarService.UnaryMethod);
+
                 Assert.Empty(activity!.Events);
             }
             finally
@@ -147,23 +149,39 @@ public class GrpcCoreServerInterceptorTests
     /// A common method to test server interceptor handler success.
     /// </summary>
     /// <param name="clientRequestFunc">The specific client request function.</param>
+    /// <param name="expectedMethodName">The expected gRPC method name.</param>
     /// <param name="additionalMetadata">The additional metadata, if any.</param>
     /// <returns>A Task.</returns>
-    private static async Task TestHandlerSuccess(Func<Foobar.FoobarClient, Metadata?, Task> clientRequestFunc, Metadata? additionalMetadata = null)
+    private static async Task TestHandlerSuccess(
+        Func<Foobar.FoobarClient, Metadata?, Task> clientRequestFunc,
+        string expectedMethodName,
+        Metadata? additionalMetadata = null)
     {
         // starts the server with the server interceptor
         var testTags = new TestActivityTags();
-        var interceptorOptions = new ServerTracingInterceptorOptions { Propagator = new TraceContextPropagator(), RecordMessageEvents = true, AdditionalTags = testTags.Tags };
+        var interceptorOptions = new ServerTracingInterceptorOptions
+        {
+            Propagator = new TraceContextPropagator(),
+            RecordMessageEvents = true,
+            AdditionalTags = testTags.Tags,
+        };
+
         using var server = FoobarService.Start(new ServerTracingInterceptor(interceptorOptions));
 
         // No parent Activity, no context from header
         using (var activityListener = new InterceptorActivityListener(testTags))
         {
-            var client = FoobarService.ConstructRpcClient(server.UriString);
+            var client = FoobarService.ConstructRpcClient(server.Target);
             await clientRequestFunc(client, additionalMetadata);
 
             var activity = activityListener.Activity;
-            GrpcCoreClientInterceptorTests.ValidateCommonActivityTags(activity, StatusCode.OK, interceptorOptions.RecordMessageEvents);
+
+            GrpcCoreClientInterceptorTests.ValidateCommonActivityTags(
+                activity,
+                expectedMethodName,
+                StatusCode.OK,
+                interceptorOptions.RecordMessageEvents);
+
             Assert.NotNull(activity);
             Assert.Equal(default, activity.ParentSpanId);
         }
@@ -172,13 +190,19 @@ public class GrpcCoreServerInterceptorTests
         using (var activityListener = new InterceptorActivityListener(testTags))
         {
             var client = FoobarService.ConstructRpcClient(
-                server.UriString,
+                server.Target,
                 additionalMetadata: [new Metadata.Entry("traceparent", FoobarService.DefaultTraceparentWithSampling)]);
 
             await clientRequestFunc(client, additionalMetadata).ConfigureAwait(false);
 
             var activity = activityListener.Activity;
-            GrpcCoreClientInterceptorTests.ValidateCommonActivityTags(activity, StatusCode.OK, interceptorOptions.RecordMessageEvents);
+
+            GrpcCoreClientInterceptorTests.ValidateCommonActivityTags(
+                activity,
+                expectedMethodName,
+                StatusCode.OK,
+                interceptorOptions.RecordMessageEvents);
+
             Assert.NotNull(activity);
             Assert.Equal(FoobarService.DefaultParentFromTraceparentHeader.SpanId, activity.ParentSpanId);
         }
@@ -188,9 +212,13 @@ public class GrpcCoreServerInterceptorTests
     /// A common method to test server interceptor handler failure.
     /// </summary>
     /// <param name="clientRequestFunc">The specific client request function.</param>
+    /// <param name="expectedMethodName">The expected gRPC method name.</param>
     /// <param name="additionalMetadata">The additional metadata, if any.</param>
     /// <returns>A Task.</returns>
-    private static async Task TestHandlerFailure(Func<Foobar.FoobarClient, Metadata?, Task> clientRequestFunc, Metadata? additionalMetadata = null)
+    private static async Task TestHandlerFailure(
+        Func<Foobar.FoobarClient, Metadata?, Task> clientRequestFunc,
+        string expectedMethodName,
+        Metadata? additionalMetadata = null)
     {
         // starts the server with the server interceptor
         var testTags = new TestActivityTags();
@@ -199,7 +227,7 @@ public class GrpcCoreServerInterceptorTests
 
         using var activityListener = new InterceptorActivityListener(testTags);
         var client = FoobarService.ConstructRpcClient(
-            server.UriString,
+            server.Target,
             additionalMetadata:
             [
                 new Metadata.Entry("traceparent", FoobarService.DefaultTraceparentWithSampling),
@@ -210,7 +238,14 @@ public class GrpcCoreServerInterceptorTests
         await Assert.ThrowsAsync<RpcException>(async () => await clientRequestFunc(client, additionalMetadata).ConfigureAwait(false));
 
         var activity = activityListener.Activity;
-        GrpcCoreClientInterceptorTests.ValidateCommonActivityTags(activity, StatusCode.ResourceExhausted, interceptorOptions.RecordMessageEvents, interceptorOptions.RecordException);
+
+        GrpcCoreClientInterceptorTests.ValidateCommonActivityTags(
+            activity,
+            expectedMethodName,
+            StatusCode.ResourceExhausted,
+            interceptorOptions.RecordMessageEvents,
+            interceptorOptions.RecordException);
+
         Assert.NotNull(activity);
         Assert.Equal(FoobarService.DefaultParentFromTraceparentHeader.SpanId, activity.ParentSpanId);
     }