+++ src/OpenTelemetry.Instrumentation.AspNet.TelemetryHttpModule/CHANGELOG.md
@@ -6,6 +6,10 @@
   created.
   ([#4307](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4307))
 
+* Fixed internal request stop instrumentation callback exception handling so
+  telemetry processing failures do not escape request stop processing.
+  ([#4344](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4344))
+
 ## 1.15.2
 
 Released 2026-Apr-21